feature(ui-ux): refactor and update testIDs and translations
